+++ react-components/src/architecture/base/utilities/primitives/Vector3ArrayUtils.ts
