+++ react-components/src/architecture/concrete/terrain/TerrainThreeView.ts
@@ -29,7 +29,7 @@ import { getColorMap } from '../../base/utilities/colors/colorMaps';
 import { GroupThreeView } from '../../base/views/GroupThreeView';
 import { CDF_TO_VIEWER_TRANSFORMATION } from '@cognite/reveal';
 import { type RegularGrid2 } from './geometry/RegularGrid2';
-import { PrimitiveUtils } from '../../base/utilities/primitives/PrimitiveUtils';
+import { PrimitiveUtils } from '../../base/utilities/primitives/Vector3ArrayUtils';
 
 const SOLID_NAME = 'Solid';
 const CONTOURS_NAME = 'Contour';
